compress long names
For example, a task name 400 chars long with object references should be
turned into a stable id, at least.  This way the cardinality of the
metric is preserved, meaning we distinguish the tasks and operators from
each other, even if the name is not particularly human readable. But the
metric name components are not going to break a typical metric backend.
Letting long invalid names flow to a backend almost guarantees that some
conversion will happen and perhaps truncate and then conflate or
otherwise drop data points.